Moving the string encoder/decoder to a separate file
